+++ vision_local_core/__init__.py
@@ -1,0 +1,1 @@
+"""Internal module layout for vision_local."""